DBM 更新到 6.1.3

+++ DBM-BlackrockFoundry/BlastFurnace.lua
@@ -1,7 +1,7 @@
 local mod	= DBM:NewMod(1154, "DBM-BlackrockFoundry", nil, 457)
 local L		= mod:GetLocalizedStrings()
 
-mod:SetRevision(("$Revision: 13278 $"):sub(12, -3))
+mod:SetRevision(("$Revision: 13394 $"):sub(12, -3))
 mod:SetCreatureID(76809, 76806)--76809 foreman feldspar, 76806 heart of the mountain, 76809 Security Guard, 76810 Furnace Engineer, 76811 Bellows Operator, 76815 Primal Elementalist, 78463 Slag Elemental, 76821 Firecaller
 mod:SetEncounterID(1690)
 mod:SetZone()
@@ -106,6 +106,7 @@ local voiceFireCaller			= mod:NewVoice("ej9659", "Tank")
 local voiceSecurityGuard		= mod:NewVoice("ej9648", "Tank")
 
 mod:AddRangeFrameOption(8)
+mod:AddBoolOption("InfoFrame")
 mod:AddSetIconOption("SetIconOnFixate", 155196, false)
 mod:AddHudMapOption("HudMapOnBomb", 155192, false)
 mod:AddDropdownOption("VFYellType", {"Countdown", "Apply"}, "Countdown", "misc")
@@ -123,7 +124,7 @@ local activeSlagGUIDS = {}
 local activePrimalGUIDS = {}
 local activePrimal = 0 -- health report variable. no sync
 local prevHealth = 100
-local yellVolatileFire2
+local yellVolatileFire2 = mod:NewFadesYell(176121, nil, true, false)
 local UnitDebuff = UnitDebuff
 
 local BombFilter, VolatileFilter
@@ -282,6 +283,9 @@ function mod:OnCombatEnd()
 	if self.Options.HudMapOnBomb then
 		DBMHudMap:Disable()
 	end
+	if self.Options.InfoFrame then
+		DBM.InfoFrame:Hide()
+	end
 	self:UnregisterShortTermEvents()
 end
 
@@ -313,7 +317,7 @@ function mod:SPELL_AURA_APPLIED(args)
 	local spellId = args.spellId
 	if args:IsSpellID(155192, 174716) then
 		local uId = DBM:GetRaidUnitId(args.destName)
-		local _, _, _, _, _, duration, expires, _, _ = UnitDebuff(uId, args.spellName)
+		local _, _, _, _, _, duration, expires = UnitDebuff(uId, args.spellName)
 		local debuffTime = expires - GetTime()
 		if self:CheckTankDistance(args.sourceGUID, 40) and self.vb.phase == 1 then--Filter Works very poorly, probably because mob not a BOSS id. usually see ALL warnings and all HUDs :\
 			warnBomb:CombinedShow(1, args.destName)
@@ -334,6 +338,10 @@ function mod:SPELL_AURA_APPLIED(args)
 		if self.Options.RangeFrame and not UnitDebuff("player", args.spellName) then
 			DBM.RangeCheck:Show(8, BombFilter, nil, nil, nil, debuffTime + 0.5)
 		end
+		if self.vb.phase == 1 and self.Options.InfoFrame and not DBM.InfoFrame:IsShown() then
+			DBM.InfoFrame:SetHeader(args.spellName)
+			DBM.InfoFrame:Show(5, "playerbaddebuff", 155192)
+		end
 	elseif spellId == 155196 then
 		if not activeSlagGUIDS[args.sourceGUID] then
 			activeSlagGUIDS[args.sourceGUID] = true
@@ -351,20 +359,28 @@ function mod:SPELL_AURA_APPLIED(args)
 				self:Schedule(40, checkSecondSlag, self)
 			elseif self.vb.slagCount == 2 then
 				self.vb.secondSlagSpawned = true
+			else
+				timerSlagElemental:Start(nil, self.vb.slagCount+1)
 			end
 			voiceSlagElemental:Play("ej9657")
 		end
-		warnFixate:CombinedShow(1, args.destName)
-		if args:IsPlayer() then
-			specWarnFixate:Show()
-		end
-		--Update icon number even if option not enabled, so recoveryable in case person with option DCs
-		if self.vb.lastSlagIcon == 6 then--1-6 should be more than enough before reset. Do not want to use skull or x since probably set on kill targets
-			self.vb.lastSlagIcon = 0
-		end
-		self.vb.lastSlagIcon = self.vb.lastSlagIcon + 1
-		if self.Options.SetIconOnFixate then
-			self:SetIcon(args.destName, self.vb.lastSlagIcon)
+		if self.vb.phase == 2 then
+			warnFixate:CombinedShow(0.5, args.destName)
+			if args:IsPlayer() then
+				specWarnFixate:Show()
+			end
+			--Update icon number even if option not enabled, so recoveryable in case person with option DCs
+			if self.vb.lastSlagIcon == 6 then--1-6 should be more than enough before reset. Do not want to use skull or x since probably set on kill targets
+				self.vb.lastSlagIcon = 0
+			end
+			self.vb.lastSlagIcon = self.vb.lastSlagIcon + 1
+			if self.Options.SetIconOnFixate then
+				self:SetIcon(args.destName, self.vb.lastSlagIcon)
+			end
+			if self.Options.InfoFrame and not DBM.InfoFrame:IsShown() then
+				DBM.InfoFrame:SetHeader(args.spellName)
+				DBM.InfoFrame:Show(5, "playerbaddebuff", 155196)
+			end
 		end
 	elseif spellId == 158345 and self:AntiSpam(10, 3) then--Might be SPELL_CAST_SUCCESS instead.
 		self.vb.shieldDown = self.vb.shieldDown + 1
@@ -504,6 +520,9 @@ function mod:UNIT_DIED(args)
 				DBM.BossHealth:Clear()
 				DBM.BossHealth:AddBoss(76806)
 			end
+			if self.Options.InfoFrame then
+				DBM.InfoFrame:Hide()
+			end
 		end
 	elseif cid == 76808 then--Regulators
 		self.vb.machinesDead = self.vb.machinesDead + 1
@@ -536,6 +555,9 @@ function mod:UNIT_DIED(args)
 			)
 			if self.Options.HudMapOnBomb then
 				DBMHudMap:Disable()
+			end
+			if self.Options.InfoFrame then
+				DBM.InfoFrame:Hide()
 			end
 		else--Only announce 1 remaining. 0 remaining not needed, because have phase2 warn. double warn no good
 			warnRegulators:Show(2 - self.vb.machinesDead)
@@ -590,9 +612,11 @@ do
 			end
 		else
 			local bossPower = UnitPower("boss1") --Get Boss Power
-			if bossPower >= 85 and not self.vb.blastWarned and totalTime > 10 then
+			if bossPower >= 85 and not self.vb.blastWarned then
 				self.vb.blastWarned = true
-				specWarnBlast:Show()
+				if totalTime > 10 then
+					specWarnBlast:Show()
+				end
 			elseif bossPower < 5 and self.vb.blastWarned then--Should catch 0, if not, at least 1-4 will fire it but then timer may be a second or so off
 				self.vb.blastWarned = false
 				timerBlastCD:Start(totalTime)
